app/requirements.txt
asn1crypto==1.2.0
beautifulsoup4==4.8.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
config==0.4.2
cryptography==2.8
cycler==0.10.0
Django==2.2.10
django-cleanup==4.0.0
django-cors-middleware==1.4.0
django-oauth-toolkit==1.2.0
enum34==1.1.6
idna==2.8
ipaddress==1.0.23
jsonfield==2.0.2
kiwisolver==1.2.0
matplotlib==3.1.2
numpy==1.18.2
oauthlib==3.1.0
Pillow==7.1.0
psycopg2==2.8.4
pycparser==2.19
pyOpenSSL==17.5.0
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.13.0
soupsieve==2.0
sqlparse==0.3.0
urllib3==1.25.7
wordcloud==1.6.0
gunicorn==19.9.0
img2pdf
PyPDF2
python-dotenv
Jinja2